Revert "합체..."
