+++ mylog/app/write/page.tsx
@@ -2,14 +2,14 @@
 
 import dynamic from 'next/dynamic';
 
-const Editor = dynamic(() => import('../component/Editor/Tui'), {
+const Tui = dynamic(() => import('../component/Editor/Tui'), {
   ssr: false,
 });
 
 export default function Page() {
   return (
     <div>
-      <Editor initValue="write your content!" />
+      <Tui initValue="write your content!" />
     </div>
   );
 }
